[Shows] Moved all shows-related components to shows folder
