+++ lib/components/artist/shows/current_shows.js
@@ -4,9 +4,9 @@
 import React from 'react-native';
 const { View, Text, StyleSheet, ListView } = React;
 
-import Headline from '../text/headline'
-import SerifText from '../text/serif'
-import OpaqueImageView from '../opaque_image_view'
+import Headline from '../../text/headline'
+import SerifText from '../../text/serif'
+import OpaqueImageView from '../../opaque_image_view'
 import ShowMetadata from './show_metadata'
