+++ lib/components/artist/shows/index.js
@@ -4,11 +4,11 @@
 import React from 'react-native';
 const { View, Text, StyleSheet, Image, ListView } = React;
 
-import Headline from '../text/headline';
-import SerifText from '../text/serif'
+import Headline from '../../text/headline';
+import SerifText from '../../text/serif'
 import CurrentShows from './current_shows'
 import PastShows from './past_shows'
-import OpaqueImageView from '../opaque_image_view'
+import OpaqueImageView from '../../opaque_image_view'
 import ShowMetadata from './show_metadata'
