+++ lib/components/artist/shows/show_metadata.js
@@ -4,7 +4,7 @@
 import React from 'react-native';
 const { View, Text, StyleSheet } = React;
 
-import SerifText from '../text/serif'
+import SerifText from '../../text/serif'
 
 export default class ShowMetadata extends React.Component {
